move traits and general impls to ecc module
