+++ halo2-ecc/src/bls12_381/hash_to_curve.rs
@@ -1,9 +1,10 @@
 //! The chip that implements `draft-irtf-cfrg-hash-to-curve-16`
 //! https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-16
 
-use std::{iter, marker::PhantomData};
+use std::marker::PhantomData;
 
-use super::utils::*;
+use crate::bigint::utils::decode_into_bn;
+use crate::ecc::hash_to_curve::{ExpandMessageChip, HashCurveExt, HashEccChip, HashInstructions};
 use crate::ecc::EcPoint;
 use crate::ff::Field;
 use crate::fields::FieldChipExt;
@@ -15,119 +16,11 @@ use crate::{
 };
 use halo2_base::gates::{GateInstructions, RangeInstructions};
 use halo2_base::halo2_proofs::halo2curves::bls12_381::{Fq2, G2};
-use halo2_base::halo2_proofs::halo2curves::CurveExt;
 use halo2_base::utils::BigPrimeField;
 use itertools::Itertools;
 use num_bigint::BigUint;
 
 use super::{Fp2Chip, Fp2Point, G2Point};
-
-pub trait HashInstructions<F: BigPrimeField> {
-    const BLOCK_SIZE: usize;
-    const DIGEST_SIZE: usize;
-
-    type ThreadBuidler;
-
-    /// Digests input using hash function and returns finilized output.
-    /// `MAX_INPUT_SIZE` is the maximum size of input that can be processed by the hash function.
-    /// `strict` flag indicates whether to perform range check on input bytes.
-    fn digest<const MAX_INPUT_SIZE: usize>(
-        &self,
-        ctx: &mut Self::ThreadBuidler,
-        input: impl Iterator<Item = QuantumCell<F>>,
-        strict: bool,
-    ) -> Result<AssignedHashResult<F>, Error>;
-}
-
-pub trait HashEccChip<F: BigPrimeField, FC: FieldChipExt<F>, C: HashCurveExt<Base = FC::FieldType>>
-where
-    FC::FieldType: crate::ff::PrimeField,
-    FC: Selectable<F, FC::FieldPoint>,
-{
-    fn field_chip(&self) -> &FC;
-
-    fn scalar_mult_bits(
-        &self,
-        ctx: &mut Context<F>,
-        p: EcPoint<F, FC::FieldPoint>,
-        bits: Vec<AssignedValue<F>>,
-        window_bits: usize,
-    ) -> EcPoint<F, FC::FieldPoint>;
-
-    fn hash_to_field<HC: HashInstructions<F, ThreadBuidler = Context<F>>, XC: ExpandMessageChip>(
-        &self,
-        ctx: &mut HC::ThreadBuidler,
-        hash_chip: &HC,
-        msg: impl Iterator<Item = QuantumCell<F>>,
-        dst: &[u8],
-    ) -> Result<[FC::FieldPoint; 2], Error>;
-
-    fn isogeny_map(
-        &self,
-        ctx: &mut Context<F>,
-        p: EcPoint<F, FC::FieldPoint>,
-    ) -> EcPoint<F, FC::FieldPoint>;
-
-    fn clear_cofactor(
-        &self,
-        ctx: &mut Context<F>,
-        p: EcPoint<F, FC::FieldPoint>,
-    ) -> EcPoint<F, FC::FieldPoint>;
-
-    fn mul_by_bls_x(
-        &self,
-        ctx: &mut Context<F>,
-        p: EcPoint<F, FC::FieldPoint>,
-    ) -> EcPoint<F, FC::FieldPoint> {
-        let bls_x_bits = (0..64)
-            .map(|i| ((C::BLS_X >> i) & 1) as u8)
-            .map(|b| ctx.load_constant(F::from(b as u64)))
-            .collect_vec();
-
-        self.scalar_mult_bits(ctx, p, bls_x_bits, 4)
-    }
-
-    fn psi(
-        &self,
-        ctx: &mut Context<F>,
-        p: EcPoint<F, FC::FieldPoint>,
-    ) -> EcPoint<F, FC::FieldPoint> {
-        // 1 / ((u+1) ^ ((q-1)/3))
-        let psi_x = self.field_chip().load_constant(ctx, C::PSI_X);
-
-        // 1 / ((u+1) ^ (p-1)/2)
-        let psi_y = self.field_chip().load_constant(ctx, C::PSI_Y);
-
-        let x_frob = self.field_chip().conjugate(ctx, p.x);
-        let y_frob = self.field_chip().conjugate(ctx, p.y);
-
-        let x = self.field_chip().mul(ctx, x_frob, psi_x.clone());
-        let y = self.field_chip().mul(ctx, y_frob, psi_y.clone());
-
-        EcPoint::new(x, y)
-    }
-
-    fn psi2(
-        &self,
-        ctx: &mut Context<F>,
-        p: EcPoint<F, FC::FieldPoint>,
-    ) -> EcPoint<F, FC::FieldPoint> {
-        // 1 / 2 ^ ((q-1)/3)
-        let psi2_x = self.field_chip().load_constant(ctx, C::PSI2_X);
-
-        let x = self.field_chip().mul(ctx, p.x, psi2_x.clone());
-        let y = self.field_chip().negate(ctx, p.y);
-
-        EcPoint::new(x, y)
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct AssignedHashResult<F: BigPrimeField> {
-    // pub input_len: AssignedValue<F>,
-    pub input_bytes: Vec<AssignedValue<F>>,
-    pub output_bytes: [AssignedValue<F>; 32],
-}
 
 #[derive(Debug)]
 pub struct HashToCurveChip<
@@ -288,8 +181,8 @@ where
     }
 }
 
-const G2_EXT_DEGREE: usize = 2;
 
+const G2_EXT_DEGREE: usize = 2;
 // L = ceil((ceil(log2(p)) + k) / 8) (see section 5 of ietf draft link above)
 const L: usize = 64;
 
@@ -350,19 +243,19 @@ impl<'chip, F: BigPrimeField> HashEccChip<F, Fp2Chip<'chip, F>, G2>
                             let mut buf = vec![zero; fq_bytes];
                             let rem = fq_bytes - 32;
                             buf[rem..].copy_from_slice(&tv[..32]);
-                            let lo = decode_into_field_be::<F, _>(
+                            let lo = decode_into_bn::<F>(
                                 ctx,
                                 gate,
-                                buf.to_vec(),
+                                buf.iter().copied().rev().collect_vec(),
                                 &fp_chip.limb_bases,
                                 fp_chip.limb_bits(),
                             );
 
                             buf[rem..].copy_from_slice(&tv[32..]);
-                            let hi = decode_into_field_be::<F, _>(
+                            let hi = decode_into_bn::<F>(
                                 ctx,
                                 gate,
-                                buf.to_vec(),
+                                buf.into_iter().rev().collect_vec(),
                                 &fp_chip.limb_bases,
                                 fp_chip.limb_bits(),
                             );
@@ -454,128 +347,6 @@ impl<'chip, F: BigPrimeField> HashEccChip<F, Fp2Chip<'chip, F>, G2>
         // Ψ²(2P) - Ψ(P) + [x²]P - [x]Ψ(P) + [x]P - 1P => [x²-x-1]P + [x-1]Ψ(P) + Ψ²(2P)
         self.sub_unequal(ctx, t3, p, false)
     }
-}
-
-pub trait ExpandMessageChip {
-    fn expand_message<F: BigPrimeField, HC: HashInstructions<F, ThreadBuidler = Context<F>>>(
-        ctx: &mut HC::ThreadBuidler,
-        hash_chip: &HC,
-        range: &impl RangeInstructions<F>,
-        msg: impl Iterator<Item = QuantumCell<F>>,
-        dst: &[u8],
-        len_in_bytes: usize,
-    ) -> Result<Vec<AssignedValue<F>>, Error>;
-}
-
-pub struct ExpandMsgXmd;
-
-impl ExpandMessageChip for ExpandMsgXmd {
-    /// Implements [section 5.3 of `draft-irtf-cfrg-hash-to-curve-16`][expand_message_xmd].
-    ///
-    /// [expand_message_xmd]: https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-hash-to-curve-16#section-5.3
-    ///
-    /// References:
-    /// - https://github.com/cfrg/draft-irtf-cfrg-hash-to-curve/blob/6ce20a1/poc/hash_to_field.py#L89
-    /// - https://github.com/paulmillr/noble-curves/blob/bf70ba9/src/abstract/hash-to-curve.ts#L63
-    /// - https://github.com/succinctlabs/telepathy-circuits/blob/d5c7771/circuits/hash_to_field.circom#L139
-    fn expand_message<F: BigPrimeField, HC: HashInstructions<F, ThreadBuidler = Context<F>>>(
-        ctx: &mut HC::ThreadBuidler,
-        hash_chip: &HC,
-        range: &impl RangeInstructions<F>,
-        msg: impl Iterator<Item = QuantumCell<F>>,
-        dst: &[u8],
-        len_in_bytes: usize,
-    ) -> Result<Vec<AssignedValue<F>>, Error> {
-        let gate = range.gate();
-
-        let zero = ctx.load_zero();
-        let one = ctx.load_constant(F::ONE);
-
-        // assign DST bytes & cache them
-        let dst_len = ctx.load_constant(F::from(dst.as_ref().len() as u64));
-        let dst_prime = dst
-            .as_ref()
-            .iter()
-            .map(|&b| ctx.load_constant(F::from(b as u64)))
-            .chain(iter::once(dst_len))
-            .collect_vec();
-
-        // padding and length strings
-        let z_pad = i2osp(0, HC::BLOCK_SIZE, |_| zero);
-        let l_i_b_str = i2osp(len_in_bytes as u128, 2, |b| ctx.load_constant(b));
-
-        let assigned_msg = msg
-            .map(|cell| match cell {
-                QuantumCell::Existing(v) => v,
-                QuantumCell::Witness(v) => ctx.load_witness(v),
-                QuantumCell::Constant(v) => ctx.load_constant(v),
-                _ => unreachable!(),
-            })
-            .collect_vec();
-
-        // compute blocks
-        let ell = (len_in_bytes as f64 / HC::DIGEST_SIZE as f64).ceil() as usize;
-        let mut b_vals = Vec::with_capacity(ell);
-        let msg_prime = z_pad
-            .into_iter()
-            .chain(assigned_msg)
-            .chain(l_i_b_str)
-            .chain(iter::once(zero))
-            .chain(dst_prime.clone())
-            .map(QuantumCell::Existing);
-
-        let b_0 = hash_chip.digest::<143>(ctx, msg_prime, false)?.output_bytes;
-
-        b_vals.insert(
-            0,
-            hash_chip
-                .digest::<77>(
-                    ctx,
-                    b_0.into_iter()
-                        .chain(iter::once(one))
-                        .chain(dst_prime.clone())
-                        .map(QuantumCell::Existing),
-                    false,
-                )?
-                .output_bytes,
-        );
-
-        for i in 1..ell {
-            let preimg = strxor(b_0, b_vals[i - 1], gate, ctx)
-                .into_iter()
-                .chain(iter::once(ctx.load_constant(F::from(i as u64 + 1))))
-                .chain(dst_prime.clone())
-                .map(QuantumCell::Existing);
-
-            b_vals.insert(i, hash_chip.digest::<77>(ctx, preimg, false)?.output_bytes);
-        }
-
-        let uniform_bytes = b_vals.into_iter().flatten().take(len_in_bytes).collect_vec();
-
-        Ok(uniform_bytes)
-    }
-}
-
-pub trait HashCurveExt: CurveExt
-where
-    Self::Base: crate::ff::PrimeField,
-{
-    type Fp: crate::ff::PrimeField + crate::ff::WithSmallOrderMulGroup<3>;
-
-    const BLS_X: u64;
-
-    const SWU_A: Self::Base;
-    const SWU_B: Self::Base;
-    const SWU_Z: Self::Base;
-
-    const ISO_XNUM: [Self::Base; 4];
-    const ISO_XDEN: [Self::Base; 3];
-    const ISO_YNUM: [Self::Base; 4];
-    const ISO_YDEN: [Self::Base; 4];
-
-    const PSI_X: Self::Base;
-    const PSI_Y: Self::Base;
-    const PSI2_X: Self::Base;
 }
 
 mod bls12_381 {
